+++ MobileAds/AdMob/Extension/UIView+Extension+Admod.swift
@@ -26,3 +26,50 @@ extension UIView {
             }
     }
 }
+
+@IBDesignable
+class DesignableGradient: UIView {
+    @IBInspectable var startGradient: UIColor = .white
+    @IBInspectable var endGradient: UIColor = .white
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.backgroundColor = UIColor().gradientColor(bounds: self.bounds, colorStart: startGradient, colorEnd: endGradient, isHorizontalMode: false)
+    }
+    
+}
+@IBDesignable
+class DesignableGradientButton: UIButton {
+    @IBInspectable var startGradient: UIColor = .white
+    @IBInspectable var endGradient: UIColor = .white
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.backgroundColor = UIColor().gradientColor(bounds: self.bounds, colorStart: startGradient, colorEnd: endGradient, isHorizontalMode: false)
+    }
+    
+}
+
+@IBDesignable
+class DesignableGradientLablel: UILabel {
+    @IBInspectable var startGradient: UIColor = .white
+    @IBInspectable var endGradient: UIColor = .white
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.backgroundColor = UIColor().gradientColor(bounds: self.bounds, colorStart: startGradient, colorEnd: endGradient, isHorizontalMode: false)
+    }
+    
+}
+
+@IBDesignable
+class DesignableGradientLablelText: UILabel {
+    @IBInspectable var startGradient: UIColor = .white
+    @IBInspectable var endGradient: UIColor = .white
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.textColor = UIColor().gradientColor(bounds: self.bounds, colorStart: startGradient, colorEnd: endGradient, isHorizontalMode: false)
+    }
+    
+}
